third_party/WebKit/LayoutTests/platform/mac/fast/canvas/canvas-text-ideographic-space-expected.txt

FAIL spanWidth == canvasWidth should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
a　b　c
